+++ miss_hit_core/config.py
@@ -261,6 +261,10 @@ STYLE_RULES = {
     "whitespace_comma" : Style_Rule(
         "Ensures there is no whitespace before a comma and whitespace after."),
 
+    "spurious_row_comma" : Style_Rule(
+        "Ensures there are no unnecessary leading or trailing commas in a"
+        " matrix or cell."),
+
     "whitespace_colon" : Style_Rule(
         "Ensures there is no whitespace around colons except if they come"
         " after a comma."),