+++ CHANGELOG.md
@@ -38,6 +38,10 @@ None known. Should be compatible with up to MATLAB 2019b.
   annotaion blocks were always OK, but multi-line annotations were
   considered continuations.
 
+* MH Style has a new rule "spurious_row_comma" which complains about
+  trailing or starting commas in matrix and cell expressions (for
+  example: `[,1,2,]`)
+
 * MH Style should now only print `[fixed]` for problems that really
   are fixed. Specifically in files that contain parse errors, we now
   no longer pretend the problems are fixed.